feat(forge): shadow the forge runtime

+++ minecraft-codev-forge/src/main/kotlin/net/msrandom/minecraftcodev/forge/mappings/inject-forge-mapping-service.kt
@@ -1,12 +1,15 @@
 package net.msrandom.minecraftcodev.forge.mappings
 
 import net.msrandom.minecraftcodev.forge.MinecraftCodevForgePlugin
-import java.net.JarURLConnection
 import java.nio.file.FileSystem
+import java.nio.file.FileSystems
+import kotlin.io.path.ExperimentalPathApi
+import kotlin.io.path.copyToRecursively
 import kotlin.io.path.deleteExisting
+import kotlin.io.path.exists
 import kotlin.io.path.notExists
-import kotlin.io.path.outputStream
 
+@OptIn(ExperimentalPathApi::class)
 fun injectForgeMappingService(fileSystem: FileSystem): Boolean {
     val serviceFile = fileSystem.getPath("META-INF", "services", "cpw.mods.modlauncher.api.INameMappingService")
 
@@ -15,19 +18,15 @@ fun injectForgeMappingService(fileSystem: FileSystem): Boolean {
     }
 
     val injectsFolder = "/forge-mapping-injects"
-    val codevInjects = MinecraftCodevForgePlugin::class.java.getResource(injectsFolder) ?: return false
+    val codevForgeFS =
+        FileSystems.getFileSystem(MinecraftCodevForgePlugin::class.java.protectionDomain.codeSource.location.toURI())
+
+    val codevInjectPath = codevForgeFS.getPath(injectsFolder)
+    if (!codevInjectPath.exists()) return false
 
     serviceFile.deleteExisting()
 
-    val jar = (codevInjects.openConnection() as JarURLConnection).jarFile
-
-    for (entry in jar.entries()) {
-        if (entry.isDirectory || !entry.name.startsWith(injectsFolder)) continue
-
-        jar.getInputStream(entry).use {
-            it.copyTo(fileSystem.getPath(entry.name.substring(injectsFolder.length)).outputStream())
-        }
-    }
+    codevInjectPath.copyToRecursively(fileSystem.rootDirectories.first(), followLinks = true, overwrite = true)
 
     return true
 }
